fix(amazonq): refactor qpanel to be less leaky
The Q panel leaks implementation everywhere so it is difficult to cleanup and recreate the panel
